Move LanguageMenu components to subfolders and add indexes
