+++ src/core/LanguageMenu/LanguageMenu/index.tsx
@@ -1,0 +1,6 @@
+export {
+  LanguageMenu,
+  LanguageMenuProps,
+  LanguageMenuItemBaseProps,
+  LanguageMenuLinkBaseProps,
+} from './LanguageMenu';
