+++ src/core/LanguageMenu/LanguageMenuItem/index.tsx
@@ -1,0 +1,1 @@
+export { LanguageMenuItem, LanguageMenuItemProps } from './LanguageMenuItem';
