+++ src/core/LanguageMenu/LanguageMenuLink/index.tsx
@@ -1,0 +1,1 @@
+export { LanguageMenuLink, LanguageMenuLinkProps } from './LanguageMenuLink';
